search result now shows user profiles

+++ Components/Main/SearchResult.js
@@ -1,21 +1,27 @@
-import React, { useEffect } from 'react';
-import { View, Image, Text, TouchableHighlight } from 'react-native';
+import React, { useEffect, useState } from 'react';
+import { View, Image, Text, TouchableHighlight, ActivityIndicator } from 'react-native';
 import { getCachedImage, cacheImages } from '../../Functions/cacheFunctions';
 import DiscoverStyles from '../../Styles/DiscoverStyles';
 import GlobalStyles from '../../Styles/GlobalStyles';
 import Styles from '../../Styles/StyleConstants';
 
-// TODO: Figure out what the yellow error is happening here
-const SearchResult = ({ drink, navigation, removable }) => {
+// This is a horizontal list item that either includes a drink object or a user object based
+// on the passed in item
+const SearchResult = ({ item, navigation, removable }) => {
+
+
+    // Only cache the drink images
     useEffect(() => {
-        cacheImages(drink.id, drink.imageURL)
-    }, [drink]);
+        if (item) {
+            cacheImages(item.imageURL, item.id);
+        }
+    }, [item]);
 
     const renderTags = () => {
         let res = '';
-        if (drink.tags) {
-            for (let i = 0; i < drink.tags.length; i++) {
-                res += drink.tags[i].name + ', '
+        if (item.tags) {
+            for (let i = 0; i < item.tags.length; i++) {
+                res += item.tags[i].name + ', '
             };
             res = res.substr(0, res.length - 2);
             return <Text style={[GlobalStyles.paragraph3, { color: Styles.GRAY }]}>{res}</Text>
@@ -23,22 +29,40 @@ const SearchResult = ({ drink, navigation, removable }) => {
         return null;
     }
 
-    return (
-        <TouchableHighlight
-            onLongPress={() => removable && handleRemove()}
-            onPress={() => navigation.navigate('DrinkDetailScreen', { drink: drink })}
-            underlayColor={'#d6d6d6'}
-        >
-            <View style={DiscoverStyles.searchContainer}>
-                <Image source={{ uri: getCachedImage(drink.id) || drink.imageURL }} style={DiscoverStyles.searchImage} />
-                <View style={{ flexDirection: 'column', justifyContent: 'center' }}>
-                    <Text style={GlobalStyles.paragraphbold1}>{drink.name}</Text>
-                    <Text style={GlobalStyles.paragraph3}>{drink.strength.label}</Text>
-                    {renderTags()}
+    if (item.authorID) {
+        return (
+            <TouchableHighlight
+                onLongPress={() => removable && handleRemove()}
+                onPress={() => navigation.navigate('DrinkDetailScreen', { drink: item })}
+                underlayColor={'#d6d6d6'}
+            >
+                <View style={DiscoverStyles.searchContainer}>
+                    <Image source={{ uri: getCachedImage(item.id) || item.imageURL }} style={DiscoverStyles.searchImage} />
+                    <View style={{ flexDirection: 'column', justifyContent: 'center' }}>
+                        <Text style={GlobalStyles.paragraphbold1}>{item.name}</Text>
+                        <Text style={GlobalStyles.paragraph3}>{item.strength.label}</Text>
+                        {renderTags()}
+                    </View>
                 </View>
-            </View>
-        </TouchableHighlight>
-    )
+            </TouchableHighlight>
+        )
+    } else {
+        return (
+            <TouchableHighlight
+                onPress={() => navigation.navigate('ProfileScreen', { user: item })}
+                underlayColor={'#d6d6d6'}
+            >
+                <View style={DiscoverStyles.searchContainer}>
+                    <Image source={{ uri: getCachedImage(item.id) || item.imageURL }} style={[DiscoverStyles.searchImage, DiscoverStyles.searchImageProfile]} />
+                    <View style={{ flexDirection: 'column', justifyContent: 'center' }}>
+                        <Text style={GlobalStyles.paragraphbold1}>{item.fName} {item.lName}</Text>
+                        <Text style={GlobalStyles.paragraph3}>{item.userName}</Text>
+                    </View>
+                </View>
+            </TouchableHighlight>
+        )
+    }
+
 }
 
 export default SearchResult;
